separated math matrix
